chore: remove all deprecated code for v3.0
Remove approximately 4,676 lines of deprecated code accumulated during v2.x lifecycle:
- 3,726 lines: Realtime Protocol 1.0 implementation (11 files)
- 950 lines: Deprecated constructors, methods, properties, error cases

- Entire `/Sources/Realtime/Deprecated/` folder (Protocol 1.0)
- `RealtimeChannelV2.subscribe()` (use `subscribeWithError()`)
- `RealtimeChannelV2.broadcast(event:)` receiving method (use `broadcastStream(event:)`)
- Type aliases: Message, Configuration, Status, Subscription
- Property: `RealtimeClientV2.subscriptions` (use `channels`)

- Extracted `ChannelEvent` enum to `Sources/Realtime/ChannelEvent.swift` (still needed by v2)

- `Sources/Auth/Deprecated.swift`
- Deprecated constructors without `logger` parameter
- `AuthAdmin.deleteUser(id: String)` (use UUID version)
- Error cases: missingExpClaim, malformedJWT, sessionNotFound, pkce(_:), invalidImplicitGrantFlowURL, missingURL, invalidRedirectScheme
- Type aliases: GoTrueClient, GoTrueMFA, GoTrueLocalStorage, GoTrueMetaSecurity, GoTrueError
- Property: `User.emailChangeToken`
- Type alias: `MFAEnrollParams` (use MFATotpEnrollParams or MFAPhoneEnrollParams)

- `Sources/PostgREST/Deprecated.swift`
- Deprecated constructors without `logger` parameter
- Filter methods: like(_:value:), in(_:value:), plfts(), phfts(), wfts(), ilike(_:value:)
- Property: `PostgrestFilterValue.queryValue` (use `rawValue`)
- Type alias: URLQueryRepresentable

- `Sources/Storage/Deprecated.swift`
- Deprecated constructors without `logger` parameter
- Upload methods with String return type (use FileUploadResponse)
- Upload methods with old parameter names (file: → data:)
- Types: File struct, FormData class
- Public access to defaultStorageEncoder and defaultStorageDecoder

- `Sources/Supabase/Deprecated.swift`
- Property: `.database` (use `.from()`, `.rpc()`, `.schema()`)
- Property: `.realtime` (use `.realtimeV2`)
- Internal `_realtime` property and initialization

- Updated test usage of deprecated APIs
- Removed `emailChangeToken` from test UserAttributes
- Changed `MFAEnrollParams` to `MFATotpEnrollParams` in tests
- Removed test usage of `defaultStorageEncoder`

Added comprehensive `MIGRATION_V3.md` with:
- Complete breaking changes list
- Migration examples for all removed APIs
- Checklist for migration
- Links to additional resources

All removed APIs were deprecated in v2.x with clear migration paths. This is a v3.0 major version bump.

Closes: SDK-436
