+++ Sources/Auth/AuthError.swift
@@ -130,71 +130,6 @@ public enum AuthError: LocalizedError, Equatable {
   @available(
     *,
     deprecated,
-    message:
-      "Error used to be thrown when no exp claim was found in JWT during setSession(accessToken:refreshToken:) method."
-  )
-  case missingExpClaim
-
-  @available(
-    *,
-    deprecated,
-    message:
-      "Error used to be thrown when provided JWT wasn't valid during setSession(accessToken:refreshToken:) method."
-  )
-  case malformedJWT
-
-  @available(*, deprecated, renamed: "sessionMissing")
-  public static var sessionNotFound: AuthError { .sessionMissing }
-
-  /// Error thrown during PKCE flow.
-  @available(
-    *,
-    deprecated,
-    renamed: "pkceGrantCodeExchange",
-    message: "Error was grouped in `pkceGrantCodeExchange`, please use it instead of `pkce`."
-  )
-  public static func pkce(_ reason: PKCEFailureReason) -> AuthError {
-    switch reason {
-    case .codeVerifierNotFound:
-      .pkceGrantCodeExchange(message: "A code verifier wasn't found in PKCE flow.")
-    case .invalidPKCEFlowURL:
-      .pkceGrantCodeExchange(message: "Not a valid PKCE flow url.")
-    }
-  }
-
-  @available(*, deprecated, message: "Use `pkceGrantCodeExchange` instead.")
-  public enum PKCEFailureReason: Sendable {
-    /// Code verifier not found in the URL.
-    case codeVerifierNotFound
-
-    /// Not a valid PKCE flow URL.
-    case invalidPKCEFlowURL
-  }
-
-  @available(*, deprecated, renamed: "implicitGrantRedirect")
-  public static var invalidImplicitGrantFlowURL: AuthError {
-    .implicitGrantRedirect(message: "Not a valid implicit grant flow url.")
-  }
-
-  @available(
-    *,
-    deprecated,
-    message:
-      "This error is never thrown, if you depend on it, you can remove the logic as it never happens."
-  )
-  case missingURL
-
-  @available(
-    *,
-    deprecated,
-    message:
-      "Error used to be thrown on methods which required a valid redirect scheme, such as signInWithOAuth. This is now considered a programming error an a assertion is triggered in case redirect scheme isn't provided."
-  )
-  case invalidRedirectScheme
-
-  @available(
-    *,
-    deprecated,
     renamed: "api(message:errorCode:underlyingData:underlyingResponse:)"
   )
   public static func api(_ error: APIError) -> AuthError {
@@ -283,11 +218,6 @@ public enum AuthError: LocalizedError, Equatable {
       .implicitGrantRedirect(let message),
       .jwtVerificationFailed(let message):
       message
-    // Deprecated cases
-    case .missingExpClaim: "Missing expiration claim in the access token."
-    case .malformedJWT: "A malformed JWT received."
-    case .invalidRedirectScheme: "Invalid redirect scheme."
-    case .missingURL: "Missing URL."
     }
   }
 
@@ -298,8 +228,6 @@ public enum AuthError: LocalizedError, Equatable {
     case .api(_, let errorCode, _, _): errorCode
     case .pkceGrantCodeExchange, .implicitGrantRedirect: .unknown
     case .jwtVerificationFailed: .invalidJWT
-    // Deprecated cases
-    case .missingExpClaim, .malformedJWT, .invalidRedirectScheme, .missingURL: .unknown
     }
   }
