+++ Sources/Helpers/EventEmitter.swift
@@ -35,7 +35,6 @@ public final class ObservationToken: @unchecked Sendable, Hashable {
     mutableState.withValue { $0.onCancel = handler }
   }
 
-  @available(*, deprecated, renamed: "cancel")
   public func remove() {
     cancel()
   }
